+++ OpenGLCourseApp/PointLight.cpp
@@ -32,13 +32,13 @@ APointLight::~APointLight()
 {}
 
 void APointLight::UseLight(
-	GLfloat ambientIntensityLocation, 
-	GLfloat ambientColorLocation, 
-	GLfloat diffuseIntensityLocation, 
-	GLfloat PositionLocation, 
-	GLfloat ConstantLocation, 
-	GLfloat LinearLocation, 
-	GLfloat ExponentLocation)
+	GLuint ambientIntensityLocation,
+	GLuint ambientColorLocation,
+	GLuint diffuseIntensityLocation,
+	GLuint PositionLocation,
+	GLuint ConstantLocation,
+	GLuint LinearLocation,
+	GLuint ExponentLocation)
 {
 	ALight::UseLight(ambientIntensityLocation, ambientColorLocation, diffuseIntensityLocation);
 	glUniform3f(PositionLocation, Position.x, Position.y, Position.z);
